for closing github issue #6 - minetest.get_node_or_nil - not always checked for nil return value #6
    The meaning of "nil" return value is that the block is not loaded where that node is. A review of all the code where nil can be returned is needed, so that appropriate action is taken:

        Sometimes just ignore the action
        Sometimes you may wish to use a voxel_manip to read the map position.

    See an example at the dev wiki https://dev.minetest.net/minetest.get_node_or_nil

Now the code supports the "ignored" string, and should not try to place
light where it found a nil block.

I don't know how to test it well since anywhere the light is added is where you are, so it's probably always loaded, and it never destroyed blocks for me in the past (even after using travelnet teleporter) (although the original mod before forking could destroy blocks, especially tops of travelnet elevators and teleporters).